refactor: move labelsdict to shared
