+++ internal/shared/labelsdict/network.go
