+++ internal/shared/labelsdict/stack.go
